+++ src/utils/index.js
@@ -11,3 +11,4 @@ export * from './xhr';
 export * from './isTestEnv';
 export * from './formValidators';
 export * from './craftLayerURL';
+export * from './loadRecords';